+++ src/main/java/dynamicquad/agilehub/dummy/DummyDataLoader.java
@@ -6,9 +6,9 @@ import dynamicquad.agilehub.dummy.bulkRepository.MemberProjectBulkRepository;
 import dynamicquad.agilehub.dummy.bulkRepository.ProjectBulkRepository;
 import dynamicquad.agilehub.issue.domain.Issue;
 import dynamicquad.agilehub.issue.domain.IssueStatus;
-import dynamicquad.agilehub.issue.domain.epic.Epic;
-import dynamicquad.agilehub.issue.domain.story.Story;
-import dynamicquad.agilehub.issue.domain.task.Task;
+import dynamicquad.agilehub.issue.domain.Epic;
+import dynamicquad.agilehub.issue.domain.Story;
+import dynamicquad.agilehub.issue.domain.Task;
 import dynamicquad.agilehub.member.domain.Member;
 import dynamicquad.agilehub.member.domain.MemberStatus;
 import dynamicquad.agilehub.project.domain.MemberProject;
