+++ src/main/java/dynamicquad/agilehub/issue/dto/backlog/EpicResponseDto.java
@@ -1,7 +1,7 @@
 package dynamicquad.agilehub.issue.dto.backlog;
 
 import dynamicquad.agilehub.issue.IssueType;
-import dynamicquad.agilehub.issue.domain.epic.Epic;
+import dynamicquad.agilehub.issue.domain.Epic;
 import dynamicquad.agilehub.member.dto.AssigneeDto;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
