+++ src/main/java/dynamicquad/agilehub/issue/service/IssueQueryService.java
@@ -4,12 +4,12 @@ import dynamicquad.agilehub.global.exception.GeneralException;
 import dynamicquad.agilehub.global.header.status.ErrorStatus;
 import dynamicquad.agilehub.issue.IssueType;
 import dynamicquad.agilehub.issue.domain.Issue;
-import dynamicquad.agilehub.issue.domain.epic.Epic;
-import dynamicquad.agilehub.issue.domain.epic.EpicRepository;
-import dynamicquad.agilehub.issue.domain.story.Story;
-import dynamicquad.agilehub.issue.domain.story.StoryRepository;
-import dynamicquad.agilehub.issue.domain.task.Task;
-import dynamicquad.agilehub.issue.domain.task.TaskRepository;
+import dynamicquad.agilehub.issue.domain.Epic;
+import dynamicquad.agilehub.issue.repository.EpicRepository;
+import dynamicquad.agilehub.issue.domain.Story;
+import dynamicquad.agilehub.issue.repository.StoryRepository;
+import dynamicquad.agilehub.issue.domain.Task;
+import dynamicquad.agilehub.issue.repository.TaskRepository;
 import dynamicquad.agilehub.issue.dto.IssueResponseDto;
 import dynamicquad.agilehub.issue.dto.SimpleIssueResponseDto;
 import dynamicquad.agilehub.issue.dto.backlog.EpicResponseDto;
