+++ src/main/java/dynamicquad/agilehub/sprint/service/SprintQueryService.java
@@ -1,7 +1,7 @@
 package dynamicquad.agilehub.sprint.service;
 
 import dynamicquad.agilehub.issue.domain.Issue;
-import dynamicquad.agilehub.issue.domain.IssueRepository;
+import dynamicquad.agilehub.issue.repository.IssueRepository;
 import dynamicquad.agilehub.member.dto.AssigneeDto;
 import dynamicquad.agilehub.project.service.ProjectQueryService;
 import dynamicquad.agilehub.sprint.controller.response.SprintResponse.IssueInSprintDto;
