+++ src/test/java/dynamicquad/agilehub/issue/comment/CommentServiceTest.java
@@ -6,7 +6,7 @@ import dynamicquad.agilehub.comment.domain.Comment;
 import dynamicquad.agilehub.comment.response.CommentResponse.CommentCreateResponse;
 import dynamicquad.agilehub.comment.service.CommentService;
 import dynamicquad.agilehub.issue.domain.IssueStatus;
-import dynamicquad.agilehub.issue.domain.epic.Epic;
+import dynamicquad.agilehub.issue.domain.Epic;
 import dynamicquad.agilehub.member.domain.Member;
 import dynamicquad.agilehub.member.dto.MemberRequestDto.AuthMember;
 import dynamicquad.agilehub.project.domain.MemberProject;
