+++ src/test/java/dynamicquad/agilehub/issue/comment/domain/CommentRepositoryTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import dynamicquad.agilehub.comment.domain.Comment;
 import dynamicquad.agilehub.comment.domain.CommentRepository;
 import dynamicquad.agilehub.issue.domain.IssueStatus;
-import dynamicquad.agilehub.issue.domain.epic.Epic;
+import dynamicquad.agilehub.issue.domain.Epic;
 import dynamicquad.agilehub.member.domain.Member;
 import dynamicquad.agilehub.project.domain.Project;
 import jakarta.persistence.EntityManager;
