+++ src/test/java/dynamicquad/agilehub/issue/domain/epic/EpicRepositoryTest.java
@@ -2,9 +2,11 @@ package dynamicquad.agilehub.issue.domain.epic;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import dynamicquad.agilehub.issue.domain.Epic;
 import dynamicquad.agilehub.issue.domain.IssueStatus;
-import dynamicquad.agilehub.issue.domain.story.Story;
+import dynamicquad.agilehub.issue.domain.Story;
 import dynamicquad.agilehub.issue.dto.backlog.EpicResponseDto;
+import dynamicquad.agilehub.issue.repository.EpicRepository;
 import dynamicquad.agilehub.project.domain.Project;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
