+++ src/test/java/dynamicquad/agilehub/sprint/SprintQueryServiceTest.java
@@ -2,9 +2,9 @@ package dynamicquad.agilehub.sprint;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import dynamicquad.agilehub.issue.domain.epic.Epic;
-import dynamicquad.agilehub.issue.domain.story.Story;
-import dynamicquad.agilehub.issue.domain.task.Task;
+import dynamicquad.agilehub.issue.domain.Epic;
+import dynamicquad.agilehub.issue.domain.Story;
+import dynamicquad.agilehub.issue.domain.Task;
 import dynamicquad.agilehub.project.domain.Project;
 import dynamicquad.agilehub.sprint.controller.response.SprintResponse.SprintReadResponse;
 import dynamicquad.agilehub.sprint.domain.Sprint;
